separate state + update engine
