+++ ignite/engine/__init__.py
@@ -6,7 +6,8 @@ import torch
 import ignite.distributed as idist
 from ignite.engine.deterministic import DeterministicEngine
 from ignite.engine.engine import Engine
-from ignite.engine.events import CallableEventWithFilter, EventEnum, Events, EventsList, RemovableEventHandle, State
+from ignite.engine.events import CallableEventWithFilter, EventEnum, Events, EventsList, RemovableEventHandle
+from ignite.engine.state import State
 from ignite.metrics import Metric
 from ignite.utils import convert_tensor
