+++ ignite/engine/events.py
@@ -2,16 +2,14 @@ import numbers
 import weakref
 from enum import Enum
 from types import DynamicClassAttribute
-from typing import TYPE_CHECKING, Any, Callable, Dict, Iterable, Iterator, List, Optional, Union
-
-from torch.utils.data import DataLoader
+from typing import TYPE_CHECKING, Any, Callable, Iterator, List, Optional, Union
 
 from ignite.engine.utils import _check_signature
 
 if TYPE_CHECKING:
-    from ignite.engine.engine import Engine
+    from ignite.base.mixins import EventsDriven
 
-__all__ = ["CallableEventWithFilter", "EventEnum", "Events", "State", "EventsList", "RemovableEventHandle"]
+__all__ = ["CallableEventWithFilter", "EventEnum", "Events", "EventsList", "RemovableEventHandle"]
 
 
 class CallableEventWithFilter:
@@ -97,7 +95,7 @@ class CallableEventWithFilter:
     def every_event_filter(every: int) -> Callable:
         """A wrapper for every event filter."""
 
-        def wrapper(engine: "Engine", event: int) -> bool:
+        def wrapper(engine: "EventsDriven", event: int) -> bool:
             if event % every == 0:
                 return True
             return False
@@ -108,7 +106,7 @@ class CallableEventWithFilter:
     def once_event_filter(once: int) -> Callable:
         """A wrapper for once event filter."""
 
-        def wrapper(engine: "Engine", event: int) -> bool:
+        def wrapper(engine: "EventsDriven", event: int) -> bool:
             if event == once:
                 return True
             return False
@@ -116,7 +114,7 @@ class CallableEventWithFilter:
         return wrapper
 
     @staticmethod
-    def default_event_filter(engine: "Engine", event: int) -> bool:
+    def default_event_filter(engine: "EventsDriven", event: int) -> bool:
         """Default event filter."""
         return True
 
@@ -348,81 +346,6 @@ class EventsList:
         return self
 
 
-class State:
-    """An object that is used to pass internal and user-defined state between event handlers. By default, state
-    contains the following attributes:
-
-    .. code-block:: python
-
-        state.iteration         # 1-based, the first iteration is 1
-        state.epoch             # 1-based, the first epoch is 1
-        state.seed              # seed to set at each epoch
-        state.dataloader        # data passed to engine
-        state.epoch_length      # optional length of an epoch
-        state.max_epochs        # number of epochs to run
-        state.max_iters         # number of iterations to run
-        state.batch             # batch passed to `process_function`
-        state.output            # output of `process_function` after a single iteration
-        state.metrics           # dictionary with defined metrics if any
-        state.times             # dictionary with total and per-epoch times fetched on
-                                # keys: Events.EPOCH_COMPLETED.name and Events.COMPLETED.name
-
-    Args:
-        kwargs: keyword arguments to be defined as State attributes.
-    """
-
-    event_to_attr = {
-        Events.GET_BATCH_STARTED: "iteration",
-        Events.GET_BATCH_COMPLETED: "iteration",
-        Events.ITERATION_STARTED: "iteration",
-        Events.ITERATION_COMPLETED: "iteration",
-        Events.EPOCH_STARTED: "epoch",
-        Events.EPOCH_COMPLETED: "epoch",
-        Events.STARTED: "epoch",
-        Events.COMPLETED: "epoch",
-    }  # type: Dict[Union[str, "Events", "CallableEventWithFilter"], str]
-
-    def __init__(self, **kwargs: Any) -> None:
-        self.iteration = 0
-        self.epoch = 0
-        self.epoch_length = None  # type: Optional[int]
-        self.max_epochs = None  # type: Optional[int]
-        self.max_iters = None  # type: Optional[int]
-        self.output = None  # type: Optional[int]
-        self.batch = None  # type: Optional[int]
-        self.metrics = {}  # type: Dict[str, Any]
-        self.dataloader = None  # type: Optional[Union[DataLoader, Iterable[Any]]]
-        self.seed = None  # type: Optional[int]
-        self.times = {
-            Events.EPOCH_COMPLETED.name: None,
-            Events.COMPLETED.name: None,
-        }  # type: Dict[str, Optional[float]]
-
-        for k, v in kwargs.items():
-            setattr(self, k, v)
-
-        self._update_attrs()
-
-    def _update_attrs(self) -> None:
-        for value in self.event_to_attr.values():
-            if not hasattr(self, value):
-                setattr(self, value, 0)
-
-    def get_event_attrib_value(self, event_name: Union[str, Events, CallableEventWithFilter]) -> int:
-        """Get the value of Event attribute with given `event_name`."""
-        if event_name not in State.event_to_attr:
-            raise RuntimeError(f"Unknown event name '{event_name}'")
-        return getattr(self, State.event_to_attr[event_name])
-
-    def __repr__(self) -> str:
-        s = "State:\n"
-        for attr, value in self.__dict__.items():
-            if not isinstance(value, (numbers.Number, str)):
-                value = type(value)
-            s += f"\t{attr}: {value}\n"
-        return s
-
-
 class RemovableEventHandle:
     """A weakref handle to remove a registered event.
 
@@ -453,7 +376,10 @@ class RemovableEventHandle:
     """
 
     def __init__(
-        self, event_name: Union[CallableEventWithFilter, Enum, EventsList, Events], handler: Callable, engine: "Engine"
+        self,
+        event_name: Union[CallableEventWithFilter, Enum, EventsList, Events],
+        handler: Callable,
+        engine: "EventsDriven",
     ) -> None:
         self.event_name = event_name
         self.handler = weakref.ref(handler)
